# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/supplier": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "List all suppliers",
                "operationId": "a0a9d26b1d95484f92178dcb634b91d2",
                "responses": {
                    "200": {
                        "description": "List of suppliers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Acme Supplier"
                                            },
                                            "contact_info": {
                                                "properties": {
                                                    "email": {
                                                        "type": "string",
                                                        "example": "acme@supplier.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+9647700000000"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Main Street, Erbil"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Create a new supplier",
                "operationId": "c82496c5acb39ca006cace3920b58c07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Supplier"
                                    },
                                    "contact_info": {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "acme@supplier.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+9647700000000"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Main Street, Erbil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/supplier/{id}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get a specific supplier",
                "operationId": "2d53dc931024916d1988fac23abfcf88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Acme Supplier"
                                        },
                                        "contact_info": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "acme@supplier.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+9647700000000"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Main Street, Erbil"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Update a supplier",
                "operationId": "e156e434d298d24a6d0eeed0e0129e51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Supplier"
                                    },
                                    "contact_info": {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "updated@supplier.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+9647700001111"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Updated Street, Erbil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete a supplier",
                "operationId": "e7f088e481f37830d2ed117b05a6d035",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}